(Scheduler) Refactor PGANScheduler to ClusterScheduler
